+++ platform/cc/Drawable.cc
@@ -11,8 +11,8 @@ public:
 
     ~SkijaDrawableImpl() {
         JNIEnv* env;
-        if (fJavaVM->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION_1_8) == JNI_OK)
-          env->DeleteWeakGlobalRef(fObject);
+        fJavaVM->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION_1_6);
+        env->DeleteWeakGlobalRef(fObject);
     }
 
     void init(JNIEnv* e, jobject o) {