+++ platform/cc/interop.cc
@@ -566,7 +566,11 @@ namespace skija {
 
         bool onFilter(jobject obj, SkPaint& paint) {
             JNIEnv *env;
+#ifdef ANDROID
+            _vm->AttachCurrentThread(&env, NULL);
+#else
             _vm->AttachCurrentThread((void **) &env, NULL);
+#endif
             jboolean result = env->CallBooleanMethod(obj, onFilterId, reinterpret_cast<jlong>(&paint));
             _vm->DetachCurrentThread();
             return result;
@@ -578,7 +582,11 @@ namespace skija {
 
         void detach(jobject obj) {
             JNIEnv *env;
+#ifdef ANDROID
+            _vm->AttachCurrentThread(&env, NULL);
+#else
             _vm->AttachCurrentThread((void **) &env, NULL);
+#endif
             env->DeleteGlobalRef(obj);
             _vm->DetachCurrentThread();
         }